+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Views/Trades/TradeCreateView.xaml.cs
@@ -1,5 +1,9 @@
 using System.Collections.ObjectModel;
+using System.Globalization;
+using System.Text.RegularExpressions;
 using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Input;
 using TD.AppInfra.DesignTime;
 using TD.AppInfra.Services;
 using TD.Models;
@@ -64,6 +68,44 @@ namespace TD.WPF.Views.Trades
             OnSourceInitialized();
         }
         partial void OnSourceInitialized();
+
+        private void DecimalTextBox_Pasting(object sender, DataObjectPastingEventArgs e)
+        {
+            if (sender is not TextBox tb)
+                return;
+            if (e.DataObject.GetDataPresent(typeof(string)))
+            {
+                var pasteText = (string)e.DataObject.GetData(typeof(string));
+                var proposed = tb.Text.Remove(tb.SelectionStart, tb.SelectionLength).Insert(tb.SelectionStart, pasteText);
+                var decimalSeparator = CultureInfo.CurrentCulture.NumberFormat.NumberDecimalSeparator;
+                if (!Regex.IsMatch(proposed, @"^-?\d*(" + Regex.Escape(decimalSeparator) + @"\d*)?$"))
+                {
+                    e.CancelCommand();
+                }
+            }
+            else
+            {
+                e.CancelCommand();
+            }
+        }
+
+        private void DecimalTextBox_PreviewTextInput(object sender, TextCompositionEventArgs e)
+        {
+            if (sender is not TextBox tb)
+                return;
+
+            var decimalSeparator = CultureInfo.CurrentCulture.NumberFormat.NumberDecimalSeparator;
+            var proposed = tb.Text.Remove(tb.SelectionStart, tb.SelectionLength).Insert(tb.SelectionStart, e.Text);
+            e.Handled = !Regex.IsMatch(proposed, @"^-?\d*(" + Regex.Escape(decimalSeparator) + @"\d*)?$");
+        }
+
+        private void IntegerTextBox_PreviewTextInput(object sender, TextCompositionEventArgs e)
+        {
+            if (sender is not TextBox tb)
+                return;
+
+            e.Handled = !char.IsDigit(e.Text, 0);
+        }
 
         private void OkButton_Click(object sender, RoutedEventArgs e)
         {